Move homepage mock query reg to more appropriate place
